parsers/teacher_info/requirements.txt
selenium==4.15.2
psycopg2-binary==2.9.9
chromedriver-autoinstaller==0.6.2
webdriver-manager==4.0.1